Implement access checking using user auth mapping
